+++ src/core/class/AdtLocalDefinitions.ts
@@ -13,7 +13,7 @@ import type {
 } from '@mcp-abap-adt/interfaces';
 import { safeErrorMessage } from '../../utils/internalUtils';
 import type { IReadOptions } from '../shared/types';
-import { AdtClass } from './AdtClass';
+import { AdtClassMemberBase } from './AdtClassMemberBase';
 import { checkClassDefinitions } from './check';
 import { updateClassDefinitions } from './includes';
 import type { IClassState } from './types';
@@ -21,7 +21,7 @@ import type { IClassState } from './types';
 // Types defined in @mcp-abap-adt/interfaces
 export type { ILocalDefinitionsConfig } from '@mcp-abap-adt/interfaces';
 
-export class AdtLocalDefinitions extends AdtClass {
+export class AdtLocalDefinitions extends AdtClassMemberBase {
   public readonly objectType: string = 'LocalDefinitions';
 
   /**
@@ -49,93 +49,6 @@ export class AdtLocalDefinitions extends AdtClass {
       validationResponse: checkResponse,
       errors: [],
     };
-  }
-
-  /**
-   * Create local definitions with full operation chain
-   * Requires parent class to be locked
-   */
-  async create(
-    config: Partial<ILocalDefinitionsConfig>,
-    options?: IAdtOperationOptions,
-  ): Promise<IClassState> {
-    if (!config.className) {
-      throw new Error('Class name is required');
-    }
-    if (!config.definitionsCode) {
-      throw new Error('Definitions code is required');
-    }
-
-    let lockHandle: string | undefined;
-    const state: IClassState = {
-      errors: [],
-    };
-
-    try {
-      // 1. Lock parent class (stateful only for lock)
-      this.logger?.info?.('Step 1: Locking parent class');
-      lockHandle = await super.lock({ className: config.className });
-      state.lockHandle = lockHandle;
-      this.logger?.info?.('Parent class locked, handle:', lockHandle);
-
-      // 2. Check local definitions code
-      const codeToCheck = options?.sourceCode || config.definitionsCode;
-      if (codeToCheck) {
-        this.logger?.info?.('Step 2: Checking local definitions code');
-        const checkResponse = await checkClassDefinitions(
-          this.connection,
-          config.className,
-          codeToCheck,
-          'inactive',
-          this.contentTypes?.sourceArtifactContentType(),
-        );
-        state.checkResult = checkResponse;
-        this.logger?.info?.('Local definitions check passed');
-      }
-
-      // 3. Update local definitions
-      this.logger?.info?.('Step 3: Creating local definitions');
-      const updateResponse = await updateClassDefinitions(
-        this.connection,
-        config.className,
-        codeToCheck as string,
-        lockHandle,
-        config.transportRequest,
-        this.contentTypes?.sourceArtifactContentType(),
-      );
-      state.updateResult = updateResponse;
-      this.logger?.info?.('Local definitions created');
-
-      // 4. Unlock parent class (obligatory stateless after unlock)
-      this.logger?.info?.('Step 4: Unlocking parent class');
-      const unlockState = await super.unlock(
-        { className: config.className },
-        lockHandle,
-      );
-      state.unlockResult = unlockState.unlockResult;
-      lockHandle = undefined;
-
-      return state;
-    } catch (error: unknown) {
-      // Cleanup on error
-      if (lockHandle) {
-        try {
-          this.logger?.warn?.('Unlocking parent class during error cleanup');
-          await super.unlock({ className: config.className }, lockHandle);
-        } catch (unlockError) {
-          this.logger?.warn?.(
-            'Failed to unlock parent class after error:',
-            safeErrorMessage(unlockError),
-          );
-        }
-      }
-
-      this.logger?.error(
-        'Create LocalDefinitions failed:',
-        safeErrorMessage(error),
-      );
-      throw error;
-    }
   }
 
   /**
